requirements.txt
groq>=0.2.0
ollama>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
